Separate latency into different file from helpers to avoid pytest issue
